webrtc_server/requirements.txt
aiohttp
aiortc
# Added for direct Google Cloud integration
google-cloud-speech
google-cloud-texttospeech
